requirements.txt
apscheduler
requests
python-dotenv
colorlog